Remove update-compile-context from locators.
The idea is that the `find' function given to the compile-lib will be the
one that determines how, given a particular type of locator, how to turn
a dependency (protocol + args) into an appropriate new locator.

In particular, this means that if we have some way of specifying relative
paths (here, I'll use 'rel'), then rel('foo/bar.arr') can end up being
context-dependent:

if the current locator is a file, then it does an appropriate relative
filesystem lookup;
if the current locator is an http URL, then it does an appropriate
relative URL retrieval;
if the current locator is using gdrive, then it looks things up there
appropriately according to how files might reference each other; and so on.

In particular, this means that the same program can be moved from local
filesystem to gdrive to an HTTP URL and mean the same thing in all three
places without having to change imports.  (This means that we'll likely
not want to use 'file' as the relative import marker; instead file(),
like gdrive() and http(), should be for setting a new "root" from which
relative links are thereby referenced.)