Add test file for testing file locators.
Refactored the file locator implementation to allow for mocking of the file
system.  Realized I was missing a few bits of functionality from the
CL.Locator type, which are mostly (but not fully, see needs-compile)
done.

Also added a use of compiler/locators/file.arr in initialize-trove so it gets
pulled in.